Refactoring. Work with modules
