+++ index.html
@@ -25,30 +25,7 @@
           <div class="todo_table_header__item">Dates</div>
           <div class="todo_table_header__item">Icons</div>
         </div>
-        <div class="todo_table_row__wraper">
-          <!--    <div class="todo_table_row">
-            <div class="todo_table_row__item">Name</div>
-            <div class="todo_table_row__item">Created</div>
-            <div class="todo_table_row__item">Category</div>
-            <div class="todo_table_row__item">Content</div>
-            <div class="todo_table_row__item">Dates</div>
-            <div class="todo_table_row__item">
-              <img
-                src="/icons/edit.png"
-                alt="edit_icon"
-                class="edit_icon"
-              /><img
-                src="/icons/archive.png"
-                alt="archive_icon"
-                class="archive_icon"
-              /><img
-                src="/icons/trash.png"
-                alt="trash_icon"
-                class="trash_icon"
-              />
-            </div>
-          </div> -->
-        </div>
+        <div class="todo_table_row__wraper"></div>
       </div>
       <div class="archived_table">
         <div class="archived_table_header">
