+++ js/script.js
@@ -1,71 +1,18 @@
-const category = ["Task", "Random thought", "Idea"];
+import { Modal } from "../services/modal.js";
+import { category, activeTask, archivedTask } from "../services/constants.js";
 
-let tableFlag = "activeTask";
-const activeTask = [
-  {
-    name: "Play associations",
-    creation_time: "September 14, 2022",
-    category: "Idea",
-    content: "Play the association game",
-    dates: "06.13.24, 06.12.24",
-  },
-  {
-    name: "Walk the dog",
-    creation_time: "September 14, 2022",
-    category: "Task",
-    content: "walk the dog",
-    dates: "07.04.23, 08.05.23",
-  },
-  {
-    name: "Try to nail to the sky",
-    creation_time: "September 14, 2022",
-    category: "Random thought",
-    content: "Try to nail to the sky, interesting ...",
-    dates: "01.10.22",
-  },
-  {
-    name: "The action is performed by the beneficiary",
-    creation_time: "September 14, 2022",
-    category: "Random thought",
-    content: "Truely-true",
-    dates: "02.11.22",
-  },
-  {
-    name: "Take out the trash",
-    creation_time: "September 14, 2022",
-    category: "Task",
-    content: "Very important!",
-    dates: "01.11.22",
-  },
-  {
-    name: "Punish an imaginary friend",
-    creation_time: "September 14, 2022",
-    category: "Random thought",
-    content: "He tired",
-    dates: "05.12.22",
-  },
-  {
-    name: "Drill some pear",
-    creation_time: "September 14, 2022",
-    category: "Task",
-    content: "Some drill",
-    dates: "12.12.22",
-  },
-];
-const archivedTask = [];
+export let tableFlag = "activeTask";
 
-const body = document.querySelector("body");
-const createNoteButton = document.querySelector(".add_task__button");
-const mainWraper = document.querySelector(".wrapper");
 const todoTable = document.querySelector(".todo_table");
 const archivedTable = document.querySelector(".archived_table");
-const modalWraper = document.querySelector(".modal_wrapper");
-
+const pivotTable = document.querySelector(".pivot_table_row__wraper");
 const todoTableWraper = document.querySelector(".todo_table_row__wraper");
 const archivedTableWraper = document.querySelector(
   ".archived_table_row__wraper"
 );
-const pivotTable = document.querySelector(".pivot_table_row__wraper");
+
+const createNoteButton = document.querySelector(".add_task__button");
+
 const switchButtonWraper = document.querySelector(".switch_button_wraper");
 
 function CreateSwitchButton() {
@@ -77,7 +24,6 @@ function CreateSwitchButton() {
   const switchButton = document.createElement("button");
   switchButton.classList.add("switch_button");
   switchButton.textContent = buttonContent;
-  console.log(switchButton);
 
   switchButtonWraper.appendChild(switchButton);
   document.querySelector(".switch_button").addEventListener("click", () => {
@@ -94,10 +40,9 @@ function CreateSwitchButton() {
 
 function FlagSwitch() {
   tableFlag = tableFlag === "activeTask" ? "archivedTask" : "activeTask";
-  console.log(tableFlag);
 }
 
-function InitTodoTable() {
+export function InitTodoTable() {
   todoTable.style.display = "";
   archivedTable.style.display = "none";
   todoTableWraper.innerHTML = "";
@@ -139,7 +84,7 @@ function InitArchiveTable() {
   });
 }
 
-function InitPivotTable() {
+export function InitPivotTable() {
   pivotTable.innerHTML = "";
 
   category.forEach((elem, idx) => {
@@ -159,45 +104,6 @@ function InitPivotTable() {
         `;
 
     pivotTable.appendChild(newElement);
-  });
-}
-
-function transformDates(dates) {
-  const regexp = /\d{1,2}\.\d{1,2}.\d{2,4}/g;
-
-  return dates.match(regexp).join(", ");
-}
-
-function CreateNoteModal(name, category, content, datesprops) {
-  const dates = transformDates(datesprops);
-  activeTask.push({
-    name,
-    creation_time: new Date().toLocaleDateString("en-US", {
-      month: "long",
-      day: "numeric",
-      year: "numeric",
-    }),
-    category,
-    content,
-    dates,
-  });
-}
-
-function EditNoteModal(
-  idx,
-  name,
-  creation_time,
-  category,
-  content,
-  datesprops
-) {
-  const dates = transformDates(datesprops);
-  activeTask.splice(idx, 1, {
-    name,
-    creation_time,
-    category,
-    content,
-    dates,
   });
 }
 
@@ -240,144 +146,6 @@ function EditNode(arr, idx) {
     task.dates,
     idx
   );
-}
-
-function Modal(
-  action,
-  inputValue = "",
-  creation_time = "",
-  inputCategory = "Task",
-  contentContent = "",
-  inputDates = "",
-  idx
-) {
-  let actionLabel;
-
-  switch (action) {
-    case "createNote":
-      actionLabel = "Add Note";
-      break;
-    case "editNote":
-      actionLabel = "Edit Note";
-      break;
-
-    default:
-      actionLabel = "?select action?";
-  }
-
-  const selectHTML = document.createElement("select");
-  selectHTML.name = "category_select";
-  selectHTML.id = "category_select";
-
-  const optionHTML = document.createElement("option");
-  optionHTML.value = inputCategory;
-  optionHTML.innerText = inputCategory;
-
-  selectHTML.appendChild(optionHTML);
-
-  category.forEach((elem, idx) => {
-    const optionHTML = document.createElement("option");
-    optionHTML.value = elem;
-    optionHTML.innerText = elem;
-
-    if (elem === inputCategory) {
-      return;
-    }
-
-    selectHTML.appendChild(optionHTML);
-  });
-
-  function CleanInputValue() {
-    (nameInput.value = ""), (contentInput.value = ""), (datesInput.value = "");
-  }
-
-  modalWraper.innerHTML = `
-
-      
-        <div class="modal_content">
-          <div class="name_input">
-            <input type="text" placeholder="Add a name" value=${inputValue}>
-          </div>
-          <div class="category_input">
-            <label for="category_select">Select a category:</label>
-      
-          </div>
-          <div class="content_input">
-            <input type="text" placeholder="Add a note content" value=${contentContent} >
-          </div>
-          <div class="dates_input">
-            <input type="text" placeholder="Add a dates" value=${inputDates} >
-          </div>
-
-          <div class="submit">
-            <button class="add_btn">${actionLabel}</button>
-          </div>
-          <div class="modal_overlay"></div>
-        </div>
-        `;
-
-  const addButton = document.querySelector(".add_btn");
-  const nameInput = document.querySelector(".name_input input");
-  const categoryInput = document.querySelector(".category_input");
-
-  const contentInput = document.querySelector(".content_input input");
-  const datesInput = document.querySelector(".dates_input input");
-  const modalOverlay = document.querySelector(".modal_overlay");
-
-  categoryInput.appendChild(selectHTML);
-  const categoryInputSelect = document.querySelector(".category_input select");
-  mainWraper.style.display = "none";
-  modalWraper.style.display = "flex";
-
-  modalOverlay.style.display = "flex";
-
-  modalOverlay.addEventListener("click", () => {
-    mainWraper.style.display = "flex";
-    modalWraper.style.display = "none";
-    modalOverlay.style.display = "none";
-  });
-
-  switch (action) {
-    case "createNote":
-      addButton.addEventListener("click", () => {
-        CreateNoteModal(
-          nameInput.value,
-          categoryInputSelect.value,
-          contentInput.value,
-          datesInput.value
-        );
-
-        CleanInputValue();
-        if (tableFlag === "activeTask") {
-          InitTodoTable();
-        }
-        InitPivotTable();
-        modalOverlay.click();
-      });
-      break;
-    case "editNote":
-      addButton.addEventListener("click", () => {
-        EditNoteModal(
-          idx,
-          nameInput.value,
-          creation_time,
-          categoryInputSelect.value,
-          contentInput.value,
-          datesInput.value
-        );
-
-        CleanInputValue();
-        if (tableFlag === "activeTask") {
-          InitTodoTable();
-        }
-        InitPivotTable();
-        modalOverlay.click();
-      });
-      break;
-
-    default:
-      break;
-  }
 }
 
 InitTodoTable();
