+++ services/constants.js
@@ -1,0 +1,56 @@
+export const category = ["Task", "Random thought", "Idea"];
+
+export const activeTask = [
+  {
+    name: "Play associations",
+    creation_time: "September 14, 2022",
+    category: "Idea",
+    content: "Play the association game",
+    dates: "06.13.24, 06.12.24",
+  },
+  {
+    name: "Walk the dog",
+    creation_time: "September 14, 2022",
+    category: "Task",
+    content: "walk the dog",
+    dates: "07.04.23, 08.05.23",
+  },
+  {
+    name: "Try to nail to the sky",
+    creation_time: "September 14, 2022",
+    category: "Random thought",
+    content: "Try to nail to the sky, interesting ...",
+    dates: "01.10.22",
+  },
+  {
+    name: "The action is performed by the beneficiary",
+    creation_time: "September 14, 2022",
+    category: "Random thought",
+    content: "Truely-true",
+    dates: "02.11.22",
+  },
+  {
+    name: "Take out the trash",
+    creation_time: "September 14, 2022",
+    category: "Task",
+    content: "Very important!",
+    dates: "01.11.22",
+  },
+  {
+    name: "Punish an imaginary friend",
+    creation_time: "September 14, 2022",
+    category: "Random thought",
+    content: "He tired",
+    dates: "05.12.22",
+  },
+  {
+    name: "Drill some pear",
+    creation_time: "September 14, 2022",
+    category: "Task",
+    content: "Some drill",
+    dates: "12.12.22",
+  },
+];
+export const archivedTask = [];
+
+export const regexp = /\d{1,2}\.\d{1,2}.\d{2,4}/g;
